Sonar Critical Fix
WorkflowResolver.java: L47
Sonar link:https://sonar.onap.org/component_issues/index?id=org.openecomp.appc%3Aappc#resolved=false|severities=CRITICAL|rules=squid%3AS2184
Issue-Id :APPC-238

+++ appc-dispatcher/appc-workflow-management/appc-workflow-management-core/src/main/java/org/openecomp/appc/workflow/impl/WorkflowResolver.java
@@ -43,7 +43,7 @@ class WorkflowResolver {
 
     private final ReentrantLock INIT_LOCK = new ReentrantLock();
 
-    WorkflowResolver(int interval) {
+    WorkflowResolver(long interval) {
         this.interval = interval * 1000;
     }
